+++ public/manifest.json
@@ -1,7 +1,7 @@
 {
     "name": "__MSG_appName__",
     "description": "__MSG_appDescription__",
-    "version": "1.0.3.1",
+    "version": "1.0.3.2",
     "manifest_version": 3,
     "default_locale": "en",
     "icons": {
